% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blindscore_ui.R
\name{blindscore_ui}
\alias{blindscore_ui}
\title{blindscore_ui}
\usage{
blindscore_ui(cdir, sheet)
}
\arguments{
\item{cdir}{The full path to the directory that contains the existing GoPro images that were previously extracted from the videos. This path should end with either '/Side A' or '/Side B'.}

\item{sheet}{The filename of the .xlsx file created using sheet_ui(). Example: 'BroadCreek_2023308_NoMaxScores.xlsx'}
}
\value{
A copied folder of the original GoPro images for the site renamed
for blind scoring and the images renamed according to their designated random
assignment number.
}
\description{
A user interface function that creates a copy of the original image folder,
renames the folder './Side A (Blind Scoring)' or './Side A (Blind Scoring)',
and renames the images according to their designated random assignment
number.
}
