% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camdir.R
\name{camdir}
\alias{camdir}
\title{camdir}
\usage{
camdir(path = "clipboard")
}
\arguments{
\item{path}{The path to the folder containing the GoPro images
to be analyzed. For the default method: path = "clipboard" attempts to use
most recent item copied to convert the toward an R readable path.}
}
\value{
The path to the folder containing the GoPro images.
}
\description{
A function that generates the path to the folder containing the GoPro images
to be analyzed in a format that R can read. For optimal efficiency, copy the
path to this folder prior to running the function to utilize the default
setting.
}
\examples{
## Pass a directory path copied from the File Explorer that uses escape keys as the standard folder-subfolder separation.
camdir(path = "C:\Desktop\GitHub\Oysters\RAP\Data")
}
