% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camside_ui.R
\name{camside_ui}
\alias{camside_ui}
\title{camside_ui}
\usage{
camside_ui()
}
\value{
The camera side: A or B.
}
\description{
A user interface function that takes user input to obtain and verify the
GoPro camera side being scored.
}
