% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasheet.R
\name{datasheet}
\alias{datasheet}
\title{datasheet}
\usage{
datasheet()
}
\value{
An excel file.
}
\description{
This function creates an excel spreadsheet from the dataframes created in
previous functions. The spreadsheet produced is used as a scoring sheet to
score GoPro images.

This function will not work on its own because it uses objects that are
created in other functions. However, it will work in sheet_ui() in
conjunction with other preceding functions.

Functions utilized from oysteRcam: tojulian().
}
