% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconstruct.R
\name{deconstruct}
\alias{deconstruct}
\title{deconstruct}
\usage{
deconstruct()
}
\value{
An excel file.
}
\description{
The function that gathers all the necessary metadata from a GoPro image's
filename.

Functions utilized from oysteRcam: camdir(), waterbody(), and fromjulian().
}
