% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromjulian.R
\name{fromjulian}
\alias{fromjulian}
\title{fromjulian}
\usage{
fromjulian(jdate, year)
}
\arguments{
\item{jdate}{Julian date as three numbers: ###.}

\item{year}{Year as four numbers: YYYY.}
}
\value{
The date in MM/DD/YYYY format converted from the Julian Date.
}
\description{
A function that takes a Julian date and a year as arguments and converts it
into the date format MM/DD/YYYY.
}
\examples{
## Pass a Julian date from a year that is NOT a leap year.
fromjulian(jdate = 289, year = 2023)

## Pass the same Julian date from a year that IS a leap year.
fromjulian(jdate = 289, year = 2020)

## Now pass a Julian date from a year that is NOT a leap year and before to March 1.
fromjulian(jdate = 43, year = 2021)

## Pass the same Julian date from a year that IS a leap year and before February 29.
fromjulian(jdate = 43, year = 2016)

}
