% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leapyear.R
\name{leapyear}
\alias{leapyear}
\title{leapyear}
\usage{
leapyear()
}
\value{
A dataframe for a leap year. The dataframe will be stored as
"calendar' and assigned to the global environment.
}
\description{
A function that creates a dataframe for a leap year with the following
variables:

num_leap: The month's two digit sequential number. Ex: January = 01.

mon_leap: The name of the month. Ex: October.

day_leap: The day in the month. Ex: July 23 = 23.

jdate_leap: The Julian date corresponding to a leap year.
}
