% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_sites_ui.R
\name{missing_sites_ui}
\alias{missing_sites_ui}
\title{missing_sites_ui}
\usage{
missing_sites_ui()
}
\value{
A vector containing the sites where videos and, consequently, images
were not recorded. This vector is assigned to the global environment.
}
\description{
A user interface function that takes user input to obtain and verify sites
that were sampled but videos were not recording, therefore, no images from
said sites are available.
}
