% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regyear.R
\name{regyear}
\alias{regyear}
\title{regyear}
\usage{
regyear()
}
\value{
A dataframe for a regular year. The dataframe will be stored as
"calendar' and assigned to the global environment.
}
\description{
A function that creates a dataframe for a regular year with the following
variables:

num_reg: The month's two digit sequential number. Ex: January = 01.

mon_reg: The name of the month. Ex: October.

day_reg: The day in the month. Ex: July 23 = 23.

jdate_reg: The Julian date corresponding to a regular year.
}
