% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_date_ui.R
\name{sampling_date_ui}
\alias{sampling_date_ui}
\title{sampling_date_ui}
\usage{
sampling_date_ui()
}
\value{
A date in MM-DD-YYYY format assigned to the global environment.
}
\description{
A user interface function that takes user input to define the date on which
sampling occurred.
}
