% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet_ui.R
\name{sheet_ui}
\alias{sheet_ui}
\title{sheet_ui}
\usage{
sheet_ui()
}
\value{
An excel file.
}
\description{
The comprehensive user interface function that utilizes other functions in
oysteRcam to have the user answers prompted questions that will produce a
datasheet. This datasheet will then be used as the scoring sheet.

Functions utilized from oysteRcam: waterbody(), sampling_date_ui(),
camside_ui(), sites_ui(), missing_sites_ui(), metadata_builder(), and
datasheet().
}
