% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites.R
\name{sites}
\alias{sites}
\title{sites}
\usage{
sites()
}
\value{
A vector of site numbers. This vector is assigned to the Global
Environment.
}
\description{
A function that iterates through each file in the GoPro image folder with the
path defined by camdir() and extracts the site numbers.
}
