% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites_ui.R
\name{sites_ui}
\alias{sites_ui}
\title{sites_ui}
\usage{
sites_ui()
}
\value{
A vector containing the sites where videos and, subsequently, images
were recorded. This vector is assigned to the global environment.
}
\description{
A user interface function that takes user input to obtain and verify sites
that were sampled and videos were recorded.
}
