% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tojulian.R
\name{tojulian}
\alias{tojulian}
\title{tojulian}
\usage{
tojulian(month, day, year)
}
\arguments{
\item{month}{Month as two numbers: ##.}

\item{day}{Day as one or two numbers: # or ##.}

\item{year}{Year as four numbers: ####.}
}
\value{
The date in MM/DD/YYYY format converted from the Julian Date.
}
\description{
A function that takes a month, day, and year as arguments and converts it
into a Julian date.
}
\examples{
## Pass a date from a year that is NOT a leap year.
tojulian(month = 10, day = 14, year = 2023)

## Pass the same date from a year that IS a leap year.
tojulian(month = 10, day = 14, year = 2020)

## Now pass a date from a year that is NOT a leap year and before to March 1.
tojulian(month = 02, day = 05, year = 2021)

## Pass the same Julian date from a year that IS a leap year and before February 29.
tojulian(month = 02, day = 05, year = 2016)
}
