% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterbody.R, R/waterbody_ui.R
\name{waterbody}
\alias{waterbody}
\title{waterbody}
\usage{
waterbody()

waterbody()
}
\value{
The location name associated with the site abbreviation identifier.
This object is assigned to the Global Environment.

A variable defining the sampling location. This variable is assigned
to the global environment.
}
\description{
A function that converts the site abbreviation identifier into the location's
name.

A user interface function that takes user input define the location where
sampling occurred.
}
