% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterbody_ui.R
\name{waterbody}
\alias{waterbody}
\title{waterbody_ui}
\usage{
waterbody()
}
\value{
A variable defining the sampling location. This variable is assigned
to the global environment.
}
\description{
A user interface function that takes user input define the location where
sampling occurred.
}
